+++ Nos/Controller/PagedNoteDataSource.swift
@@ -17,7 +17,8 @@ class PagedNoteDataSource<Header: View, EmptyPlaceholder: View>: NSObject, UICol
     private var pager: PagedRelaySubscription?
     private var context: NSManagedObjectContext
     private var header: () -> Header
-    private var emptyPlaceholder: () -> EmptyPlaceholder
+    private var emptyPlaceholder: (@escaping () -> Void) -> EmptyPlaceholder
+    private var onRefresh: () -> NSFetchRequest<Event>
     let pageSize = 20
     
     // We intentionally generate unique IDs for cell reuse to get around 
@@ -32,7 +33,8 @@ class PagedNoteDataSource<Header: View, EmptyPlaceholder: View>: NSObject, UICol
         collectionView: UICollectionView, 
         context: NSManagedObjectContext,
         @ViewBuilder header: @escaping () -> Header,
-        @ViewBuilder emptyPlaceholder: @escaping () -> EmptyPlaceholder
+        @ViewBuilder emptyPlaceholder: @escaping (@escaping () -> Void) -> EmptyPlaceholder,
+        onRefresh: @escaping () -> NSFetchRequest<Event>
     ) {
         self.databaseFilter = databaseFilter
         self.fetchedResultsController = NSFetchedResultsController<Event>(
@@ -47,7 +49,8 @@ class PagedNoteDataSource<Header: View, EmptyPlaceholder: View>: NSObject, UICol
         self.relay = relay
         self.header = header
         self.emptyPlaceholder = emptyPlaceholder
-        
+        self.onRefresh = onRefresh
+
         super.init()
         
         collectionView.register(
@@ -127,7 +130,13 @@ class PagedNoteDataSource<Header: View, EmptyPlaceholder: View>: NSObject, UICol
         let cell = collectionView.dequeueReusableCell(withReuseIdentifier: cellReuseID, for: indexPath) 
         
         cell.contentConfiguration = UIHostingConfiguration { 
-            NoteButton(note: note, hideOutOfNetwork: false, displayRootMessage: true)
+            NoteButton(
+                note: note,
+                hideOutOfNetwork: false,
+                repliesDisplayType: .discussion,
+                fetchReplies: true,
+                displayRootMessage: true
+            )
         }
         .margins(.horizontal, 0)
 
@@ -175,7 +184,24 @@ class PagedNoteDataSource<Header: View, EmptyPlaceholder: View>: NSObject, UICol
             
             footer.contentConfiguration = UIHostingConfiguration { 
                 if self.fetchedResultsController.fetchedObjects?.isEmpty == true {
-                    self.emptyPlaceholder()
+                    self.emptyPlaceholder { [weak collectionView] in
+                        let refreshControl = collectionView?.refreshControl
+                        if let refreshControl {
+                            collectionView?.scrollRectToVisible(
+                                refreshControl.frame,
+                                animated: true
+                            )
+                            refreshControl.beginRefreshing()
+                        }
+                        self.updateFetchRequest(self.onRefresh())
+                        collectionView?.reloadData()
+                        if let refreshControl {
+                            // Dismiss the refresh control
+                            DispatchQueue.main.asyncAfter(deadline: .now() + .seconds(1)) {
+                                refreshControl.endRefreshing()
+                            }
+                        }
+                    }
                 }
             }
             .margins(.horizontal, 0)