+++ Nos/Service/Relay/Filter.swift
@@ -4,17 +4,59 @@ import Foundation
 /// See [NIP-01](https://github.com/nostr-protocol/nips/blob/master/01.md#communication-between-clients-and-relays).
 struct Filter: Hashable, Identifiable {
     
+    /// List of author identifiers the Filter should be constrained to.
     var authorKeys: [RawAuthorID]
+
+    /// List of event identifiers the Filter should be constrained to.
     let eventIDs: [RawEventID]
+
+    /// List of Note kinds to filter
     var kinds: [EventKind]
+
+    /// Ask the relay to return events mentioned in the `tags` field.
     let eTags: [RawEventID]
+
+    /// Ask the relay to return authors mentioned in the `tags` field.
     let pTags: [RawAuthorID]
+
+    /// Query string to use in the search parameter.
     let search: String?
-    let inNetwork: Bool
+
+    /// Maximum number of items the relay can return.
     var limit: Int?
+
+    /// Ask the relay to return notes posted after a given date.
     var since: Date?
+
+    /// Ask relays to return notes posted before a given date.
     var until: Date?
-    
+
+    /// Whether the subscription should remain open listening for content
+    /// updates or should close after receiving a response from the relay.
+    var keepSubscriptionOpen: Bool
+
+    /// Initializes a Filter object
+    ///
+    /// - Parameter authorKeys: List of author identifiers the Filter should be
+    /// constrained to. Defaults to `[]`.
+    /// - Parameter eventIDs: List of event identifiers the Filter should be
+    /// constrained to. Defaults to `[]`.
+    /// - Parameter kinds: List of Note kinds to filter. Defaults to `[]`.
+    /// - Parameter eTags: Ask the relay to return events mentioned in the
+    /// `tags` field. Defaults to `[]`.
+    /// - Parameter pTags: Ask the relay to return authors mentioned in the
+    /// `tags` field. Defaults to `[]`.
+    /// - Parameter search: List of author identifiers the Filter should be
+    /// constrained to. Defaults to `nil`.
+    /// - Parameter limit: Maximum number of items the relay can return.
+    /// Defaults to `nil` (no limit).
+    /// - Parameter since: Ask the relay to return notes posted after a
+    /// given date. Defaults to `nil`.
+    /// - Parameter until: Ask relays to return notes posted before a given
+    /// date. Defaults to `nil`.
+    /// - Parameter keepSubscriptionOpen: Whether the subscription should remain
+    /// open listening for content updates or should close after receiving a
+    /// response from the relay. Defaults to `false`.
     init(
         authorKeys: [RawAuthorID] = [],
         eventIDs: [RawEventID] = [],
@@ -22,10 +64,10 @@ struct Filter: Hashable, Identifiable {
         eTags: [RawEventID] = [],
         pTags: [RawAuthorID] = [],
         search: String? = nil,
-        inNetwork: Bool = false,
         limit: Int? = nil,
         since: Date? = nil,
-        until: Date? = nil
+        until: Date? = nil,
+        keepSubscriptionOpen: Bool = false
     ) {
         self.authorKeys = authorKeys.sorted()
         self.eventIDs = eventIDs
@@ -33,10 +75,10 @@ struct Filter: Hashable, Identifiable {
         self.eTags = eTags
         self.pTags = pTags
         self.search = search
-        self.inNetwork = inNetwork
         self.limit = limit
         self.since = since
         self.until = until
+        self.keepSubscriptionOpen = keepSubscriptionOpen
     }
     
     var dictionary: [String: Any] {
@@ -95,26 +137,10 @@ struct Filter: Hashable, Identifiable {
         hasher.combine(search)
         hasher.combine(since)
         hasher.combine(until)
-        hasher.combine(inNetwork)
+        hasher.combine(keepSubscriptionOpen)
     }
     
     var id: String {
-        let intermediate: [String] = [
-            authorKeys.joined(separator: ","),
-            eventIDs.joined(separator: ","),
-            kinds.map { String($0.rawValue) }.joined(separator: ","),
-            limit?.description ?? "nil",
-            eTags.joined(separator: ","),
-            pTags.joined(separator: ","),
-            search ?? "nil",
-            since?.timeIntervalSince1970.description ?? "nil",
-            until?.timeIntervalSince1970.description ?? "nil",
-            inNetwork.description,
-        ]
-        
-        return intermediate
-            .joined(separator: "|")
-            .data(using: .utf8)!
-            .sha256
+        hashValue.description
     }
 }