+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Added a filter button to the Home tab that let's you browse all notes on a specific relay.
+- Add a Tap to Refresh button in empty profiles.
+- Update the reply count shown below each Note in a Feed.
 - Removed follower count from profile screen.
 - Fixed deep linking to profiles and notes.
 - Fixed issue where some nostr:nprofile references did not appear as links.